+++ packages/common/src/utils/index.ts
@@ -10,3 +10,5 @@ export * from "./slugify";
 export * from "./without-by-prop";
 export * from "./propify-string";
 export * from "./increment-string";
+export * from "./logger";
+export * from "./get-query-params";