+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased][HEAD]
+* New Feature
+  * **hub-common**: Added a logger wrapper class that only logs when a truthy (true, 1) debug query param is included in the url. Also added a util to parse a window's query params into a hash [130847](https://esriarlington.tpondemand.com/entity/130847-add-a-hub-logger-wrapper-method)
 
 ## [4.2.2] - June 1st 2020